Fix : fix some err in test
- staff deleteByCno에서 자꾸 외래키 참조 존재 오류가 남. 해결x

+++ src/test/java/com/evaluation/persistence/StaffRepositoryTests.java
@@ -133,9 +133,14 @@ public class StaffRepositoryTests {
 
 	@Test
 	public void testDeleteByCno() {
+
 		relationSurveyRepo.deleteAll();
+
 		relationMboRepo.deleteAll();
-		repo.deleteByCno(1L);
+
+		if ((relationSurveyRepo.findAll() == null) && (relationMboRepo.findAll() == null)) {
+			repo.deleteByCno(1L);
+		}
 	}
 
 	@Test
